ARM: dts: Add i2cmux-pinctrl config to Raspberry Pi BSC0
The first i2c controller (BSC0) on the Raspberry Pi family
can be pinmuxed to either GPIOs 0&1, 28&29, or 44&45.
The default has always been 0&1 as those are exposed on
the 40 pin GPIO header of the more recent models.
GPIOs 28&29 on Pi 0/A+/B+/2 and GPIOs 44&45 on Pi3 are
used for i2c to the DSI display and the camera module.

The i2c-mux-pinctrl driver allows pinctrl setup to be used
to expose the GPIO sets as alternate i2c adapters. This
patch configures that driver to expose two adapters -
i2c0 on GPIOs 0&1, and i2c_csi_dsi on the correct GPIOs for
the camera and display.
(This could be extended to expose all 3 options on the CM
and CM3, but there isn't a pressing case for that at present).

The one quirk with the i2cmux-pinctrl driver is that the
parent controller still exists as an i2c adapter which
will be connnected to whichever pins were last requested
(or an optional idle setting). Without additional configuration
that adapter appears first so would be numbered as adapter 0,
The child ports would end up as adapters 3&4. These have been
renumbered through using DT aliases to retain the ordering of
adapters.

The rev1 model A and B platforms used GPIOs 2&3 (BSC1) for the
camera, whilst rev2 model A and B used 0&1 (no I2C was routed
to the display). Neither exposed any of the alternate pins
that BSC0 can be routed to in a useful manner. For those
platforms the mux is not created, and BSC0 is routed to
GPIOs 0&1 as before.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>