serial: 8250: bcm2835aux: add notifier to follow clock changes
The serial port's clock is attached to RPi's vpu clock. Which might be
changed by the cpufreq driver anytime. This adds a cpufreq notifier,
which stops the serial port before the change, reconfigures it and
enables it again after the clock changes. Making sure the serial output
is sane anytime.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>