cpufreq: add driver for Raspbery Pi
Raspberry Pi's firmware offers and interface though which update it's
performance requirements. It allows us to request for specific runtime
frequencies, which the firmware might or might not respect, depending on
the firmware configuration and thermals.

A whole set of plls might be reconfigured by this interface. These
clocks, which are marked as non cached in their respective drivers, are
refreshed after every frequency change request in order to make sure all
drivers get properly notified with the clock changes.

This adds a new cpufreq driver that extracts it's frequency tables from
the device tree opp tables and operates the cpu frequency through the
firmware interface.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>