input: edt-ft5x06 - add polled input support
Some hardware configurations might pass on providing an interrupt line.
In that case there is always the option to use a polled input approach.
This patch adapts the driver for it.

The polled approach is only triggered if no interrupt is provided by the
firmware or platform data.

The patch also updates the device tree bindings to reflect the new
option.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>