arm64: mm: use ZONE_DMA instead of ZONE_DMA32
Current ZONE_DMA32 usage seems to be heavily rooted on having a 32 bit
DMA mask, which will no longer be the case on arm64 if we want to
support the RPi 4, as this board can only access the first GB of memory.

By using ZONE_DMA we can set a zone size smaller than 4GB, regardless of
the total memory size, and still have dma-direct and other related
subsystems behave correctly.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>