OF & ACPI/IORT: Use roundup_pow_of_two() when calculating masks
As roundup_pow_of_two() now returns 64bit values, we can use it to round
DMA limits incoming from firmware. This simplifies the mask calculation
code, as it abstracts the math behind it.

Apply the change to both subsystems as we try to keep the divergence as
little as possible.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>
Acked-by: Rob Herring <robh@kernel.org>