ARM: dts: bcm283x: Unify CMA configuration
With the introduction of the Raspberry Pi 4 we were forced to explicitly
configure CMA's location, since arm64 defaults it into the ZONE_DMA32
memory area, which is not good enough to perform DMA operations on that
device. To bypass this limitation a dedicated CMA DT node was created,
explicitly indicating the acceptable memory range and size.

That said, compatibility between boards is a must on the Raspberry Pi
ecosystem so this creates the CMA DT node for all the device family so
as for DT overlays to be able to update CMA's properties regardless of
the board being used.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>