staging: vchiq: Make copy_callback service specific
A copy_callback function pointer is passed into vchiq_queue_message() as
a means for the core code to be able to copy the input data into vc4's
slot memory.  For most users this implies a simple memcpy() but it might
also imply walking/copying arrays of user-space data pointers.

This is something we can't avoid if we want to limit the number of
copies done on incoming data. But the callback is service
specific, as opposed to message specific. So move it into the service
initialization, which simplifies the message queueing code path.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>