mfd: bcm2835-pm: Add support for BCM2711
In BCM2711 the new ARGON ASB took over V3D. The old ASB is still present
with the ISP and H264 bits, and V3D is in the same place in the new ASB
as the old one.

As per the devicetree bindings, BCM2711 will provide both the old and
new ASB registers, so get both of them and pass them into
'bcm2835-power,' which will take care of selecting which one to use
accordingly.

We're being extra careful when dealing with older firmware. We can't
assume all V3D users will use the new "brcm,bcm2711-pm" compatible
string, so use the fact that a 3rd resource is present in firmware as a
hint we're running on BCM2711.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>