KVM: x86/mmu, hyper-v: add support for SynIC GPA intercept messages
When we EPT fault in lower a VTL due to violation of memory access rules
setup by a higer VTL, the hypervisor needs to inform the higher VTL of
this event by means of a SynIC intercept message. This patch adds
support for doing this.

Although TLFS allows this intercept to be delivered to any vcpu, we will
deliver the intercept in the faulting vcpu and allow it to enter guest
(in the higher VTL) as we don't have anything to do untill the page
faulting condition is dealt with by the higher VTL.

Signed-off-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Signed-off-by: Evgeny Iakovlev <eyakovl@amazon.de>
Reviewed-by: Anel Orazgaliyeva <anelkz@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>