KVM: x86: Support per-apic timers
Hyper-V per-VTL apics still have their local timers active even if apic
is not effective on a given vcpu. To support this make sure that apic
timer callback paths only work with the actual apic that had its timer
fired and not assume that it is the effective one.

The actual handling of ineffective apic timers will come later when we
will deal with hyper-v specifics of it.

This change also cleans up existing code that is already referencing
correct apic but makes it a bit more obvious.

Signed-off-by: Evgeny Iakovlev <eyakovl@amazon.de>
Reviewed-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Reviewed-by: Anel Orazgaliyeva <anelkz@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>