KVM: x86: Assert a conflict between synic and hw-accel apics
Currently inactive apic timers will only fire for VSM-enabled VMs, which
means SynIC, which in turn means disabled apicv/avic. And all that means
that is we see an apic timer expiring and KVM using posted interrupts,
then it means we can't have SynIC and thus VSM enabled.

Assert that expectation with a BUG_ON and do it in a separate patch so
that we can potentially revert it when/if we can make synic compatible
with apicv.

Signed-off-by: Evgeny Iakovlev <eyakovl@amazon.de>
Reviewed-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Reviewed-by: Anel Orazgaliyeva <anelkz@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>