KVM: x86: Handle per-VTL apic events
According to TLFS when a local interrupt source fires for a given
per-VTL apic we need to:
* Switch into apic's VTL if it has higher priority than an active one
  and inject local interrupt.
* Delay injecting local interrupt until guest volunteers to return back
  to a lower priority apic VTL.

To implement this we are going to track which per-vcpu VTLs have their
apics requiring attention when particular apic accepts an irq. We will
only do that for DM_LOWEST and DM_FIXED sources, which should cover all
possible event sources (timers mostly).

Tracking that via __apic_accept_irq allows us to cover all interrupt
sources we care about: LVTT and stimers. Not other interrupt sources are
capable of delivering IRQs to per-VTL apics, but that may change.

When handling KVM_REQ_EVENT we will first force a switch into a
higher priority VTL if needed. This will happen after we had the chance
to handle potential VTL call/return hypercalls. If the enforced VTL
switch will override a previous explicit VTL call or return hypercall,
VTL switching machinery is already properly storing just-switched and
now overriden VTL context and we will come back to it later.

If higher-priority apic event fires after we've had a chance to check
KVM_REQ_EVENT but before we've re-entered the guest, standard
kvm_make_request machinery will cancel any injected events from current
apic, abort guest reentry and let us handle KVM_REQ_EVENTs again.

If we see an event for a lower-priority apic we will instead leave it
pending. When guest voluntarily returns to that VTL as a part of a
VTL switch, we will switch back to associated apic and re-inject all
events that were ready for it so they can be normally handled before
re-entering the guest.

Signed-off-by: Evgeny Iakovlev <eyakovl@amazon.de>
Reviewed-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Reviewed-by: Anel Orazgaliyeva <anelkz@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>