KVM: Add vcpu variants for guest memory acessors
All guest memory access via struct kvm pointer defaults to using the
address space ID 0. This was okay because the only consumer of address
spaces is x86 (SMM). In a future patch, we plan to use address spaces to
achieve isolation of memory between different VTLs (when VSM is enabled).

In preparation for this change, introduce vcpu variants for
kvm_{read,write}_guest_*_cached and kvm_is_visible_gfn methods.

Signed-off-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Reviewed-by: Evgeny Iakovlev <eyakovl@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>
[ Add gpc vcpu variants ]
Signed-off-by: Anel Orazgaliyeva <anelkz@amazon.de>