KVM: x86: hyper-v: Fix SynIC guest memory access
Maintaining the invariant that, guest memory writes using the vCPU
pointer are always made in the current address space context that the
guest was operating in is really useful. SynIC message delivery violates
it because it has to write to the SynIC message page from timer
callbacks which can arrive at any time regardless of the vCPU's current
address space context.

To fix this we resort to couple of approaches: for intercepts, since we
know these are going to target a higher VTL, we can force a VTL switch
to target VTL before attempting to deliver the message. For SynIC timers
on the other hand, we cannot force a direct switch because lower VTL
timers have to remain pending when higher VTL is executing. So we resort
to a deferred-write approach where we set the IRQ in the target VTL's
APIC and wait for the existing machinery in vcpu_run to kick us into
target VTL and then write the message current address space context.

Signed-off-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>