KVM: x86: hyper-v: Implement StartVirtualProcessor hypercall
StartVirtualProcessor hypercall can be used by a guest to start an AP
with a desired register state; it is functionally equivalent to
traditional INIT-SIPI-SIPI method. Counter-intuitively, its input
specifies a VTL and initial VP context, but expects the AP has to be
started in the highest enabled VTL.

This hypercall depends on GetVpIndexFromApicId hypercal to get the VP
index to be started, so this patch also implements it.

Signed-off-by: Evgeny Iakovlev <eyakovl@amazon.de>
Signed-off-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Reviewed-by: Anel Orazgaliyeva <anelkz@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>