KVM: x86: introduce support for non-executable memory regions
Add new "no exec" memory region flag and advertise it under the
new capability.

Enforce the flag by making __gfn_to_hva_many translation fail when the
region has the "no exec" flag and the attempted access is to execute,
similarly to how the "readonly" flag is enforced. When attempted access
is not to execute, the executable property is cached so that the EPT
entries have the correct permissions.

Signed-off-by: Anel Orazgaliyeva <anelkz@amazon.de>
Reviewed-by: Sebastian Ott <sebott@amazon.de>
Reviewed-by: Siddharth Chandrasekaran <sidcha@amazon.de>