KVM: x86: hyper-v, mmu: rework VSM intercept delivery
In preparation for handling non-executable and non-readable protections
faults, rework how we are detecting and handling VSM memory access
violations currently. Introduce a new bit for memory slot API that
userspace will set to indicate that this slot describes a VSM protection
modification within its address space: KVM_MEM_VSM_PROTECTED.

MMU code will now handle protection violations in a single place, in
direct_page_fault. Previously we tried to catch violations in places
where MMU page faults are resolved and a new spte is created and mapped.

Now we allow spte to be created instead and access to be retried,
only to be caught again on a subsequent EPT violation. We are then
checking if this violation refers to a GFN that belongs to a slot with
KVM_MEM_VSM_PROTECTED set, and if so, we resolve this page fault with a
RET_PF_NOACCESS. This allows kvm_mmu_page_fault to be a single decision
making point instead of 2-3 places where KVM might resolve a page fault.

Having to take an additional EPT violation is not a problem. VSM access
violations are an unhappy path that will most likely lead to a VM
termination anyway, so taking an extra exit is not a big deal.

Signed-off-by: Evgeny Iakovlev <eyakovl@amazon.de>