KVM: x86: Clear any pending events when switching VTLs
When we switch VTLs, we save all pending interrupt information into the
previous VTL context. Then, we move to the new context. However, if
there was any event (NMI, IRQ, exception) pending, we don't clear this
state for the new VTL that is to come.

Add clearing logic for any events that could still be pending. We will
reinitiate them when we return to the other VTL.

Signed-off-by: Alexander Graf <graf@amazon.com>
Reviewed-by: Nicolas Saenz Julienne <nsaenz@amazon.es>
Reviewed-by: Sebastian Ott <sebott@amazon.de>
Reviewed-by: Filippo Sironi <sironi@amazon.de>