KVM: x86: Save RAX/RCX on VTL context switch
When a VTL wants to return to the exact same state it found before it
was called - for example when it handles an interrupt - it needs to know
the state of all registers. Most registers are just kept in their
respective locations, but RAX and RCX get clobbered as part of the
calling convention and thus restored on VTL_RETURN. To give VTL1 a
chance to restore VTL0 RAX and RCX values on return from interrupt
context, let's put the original values in the shared page on switch.

This fixes a page fault (BugCheck 0xa) that we saw due to incorrect
restoration of register values during higher VTL interrupt handling.

Signed-off-by: Alexander Graf <graf@amazon.com>
Reviewed-by: Sebastian Ott <sebott@amazon.de>
Reviewed-by: Filippo Sironi <sironi@amazon.de>