KVM: x86/mmu: Introduce exception flag for unmapped GPAs
Introduce a flag in x86_exception which signals that a page walk failed
because a page table GPA wasn't backed by a memslot. This only applies to
page tables; the final physical address is not checked.

This extra flag is needed, because the normal page fault error code does
not contain a bit to signal this kind of fault.

Used in subsequent patches to give userspace information about translation
failure.

Signed-off-by: Nikolas Wipper <nikwip@amazon.de>