KVM: x86: selftests: Introduce memory attributes side-channel tests
Introduce memory attributes selftests to catch any vulnerable
side-channels.

Memory attributes access restrictions are vulnerable to side-channel
attacks. This means that any KVM operation initiated by the guest that
requires guest memory access (which is the case for most
para-virtualised interfaces) needs to consider memory attributes.

The tests confirm this requirement is upheld for a variety of use-cases,
and were selected to exercise specific approaches to guest memory
access, including:
 - kvm_read/write_guest()
 - gfn_to_hva_cache
 - gfn_to_pfn_cache
 - Guest page walker (which uses __try_cmpxchg_user())

Signed-off-by: Nicolas Saenz Julienne <nsaenz@amazon.com>